+++ poreana/diffusion.py
@@ -352,7 +352,6 @@ def mean(diff_data, dens_data, ax_area=[0.2, 0.8], is_print=True):
             if diff_bins[1]>=len(diff_y):
                 diff_bins[1] = diff_bins[0]
 
-
             # Check boundary
             if diff_bins[1] <= len(diff_x):
                 interp_x = [diff_x[bin_id] for bin_id in diff_bins]
@@ -467,7 +466,7 @@ def mc_trans_mat(link, step, kwargs={}, is_norm=False, is_diagonal=False, is_len
     plt.title("Lagtime: "+ str(step * frame_length * 10**12) + " ps", fontsize=10)
 
     # Plot the normalized transition matrix in a heatmap
-    sns.heatmap(data = trans_mat, **kwargs)
+    sns.heatmap(data=trans_mat, **kwargs)
     if is_length:
         plt.xlabel("Box length (nm)")